backend/requirements.txt
fastapi[all]==0.104.1
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
prometheus-client==0.19.0
